requirements.txt
requests
selenium
pillow
pytesseract